Add tests for new API
