+++ counterparty-lib/counterpartylib/test/complex_unit_test.py
@@ -1,4 +1,4 @@
-import pprint  # noqa: F401
+import json
 import tempfile
 
 import pytest
@@ -12,7 +12,7 @@ from counterpartylib.test import (
     conftest,  # noqa: F401
     util_test,
 )
-from counterpartylib.test.fixtures.params import ADDR, DP  # noqa: F401
+from counterpartylib.test.fixtures.params import ADDR
 from counterpartylib.test.util_test import CURR_DIR
 
 FIXTURE_SQL_FILE = CURR_DIR + "/fixtures/scenarios/unittest_fixture.sql"
@@ -330,238 +330,6 @@ def test_updated_tables_endpoints():
             }
 
 
-""" 
-@pytest.mark.usefixtures("api_server")
-def test_new_get_balances_by_address():
-    alice = ADDR[0]
-    url = f"{config.API_ROOT}/addresses/{alice}/balances"
-    result = requests.get(url)  # noqa: S113
-    assert result.json() == [
-        {
-            "address": "mn6q3dS2EnDUx3bmyWc6D4szJNVGtaR7zc",
-            "asset": "A95428956661682277",
-            "quantity": 100000000,
-        },
-        {
-            "address": "mn6q3dS2EnDUx3bmyWc6D4szJNVGtaR7zc",
-            "asset": "CALLABLE",
-            "quantity": 1000,
-        },
-        {
-            "address": "mn6q3dS2EnDUx3bmyWc6D4szJNVGtaR7zc",
-            "asset": "DIVISIBLE",
-            "quantity": 98800000000,
-        },
-        {
-            "address": "mn6q3dS2EnDUx3bmyWc6D4szJNVGtaR7zc",
-            "asset": "LOCKED",
-            "quantity": 1000,
-        },
-        {
-            "address": "mn6q3dS2EnDUx3bmyWc6D4szJNVGtaR7zc",
-            "asset": "MAXI",
-            "quantity": 9223372036854775807,
-        },
-        {
-            "address": "mn6q3dS2EnDUx3bmyWc6D4szJNVGtaR7zc",
-            "asset": "NODIVISIBLE",
-            "quantity": 985,
-        },
-        {
-            "address": "mn6q3dS2EnDUx3bmyWc6D4szJNVGtaR7zc",
-            "asset": "PARENT",
-            "quantity": 100000000,
-        },
-        {
-            "address": "mn6q3dS2EnDUx3bmyWc6D4szJNVGtaR7zc",
-            "asset": "XCP",
-            "quantity": 91875000000,
-        },
-    ]
-
-
-@pytest.mark.usefixtures("api_server")
-def test_new_get_balances_by_asset():
-    asset = "XCP"
-    url = f"{config.API_ROOT}/assets/{asset}/balances"
-    result = requests.get(url)  # noqa: S113
-    assert result.json() == [
-        {
-            "address": "1_mn6q3dS2EnDUx3bmyWc6D4szJNVGtaR7zc_mtQheFaSfWELRB2MyMBaiWjdDm6ux9Ezns_2",
-            "asset": "XCP",
-            "quantity": 300000000,
-        },
-        {
-            "address": "2MyJHMUenMWonC35Yi6PHC7i2tkS7PuomCy",
-            "asset": "XCP",
-            "quantity": 46449548498,
-        },
-        {
-            "address": "mn6q3dS2EnDUx3bmyWc6D4szJNVGtaR7zc",
-            "asset": "XCP",
-            "quantity": 91875000000,
-        },
-        {
-            "address": "mqPCfvqTfYctXMUfmniXeG2nyaN8w6tPmj",
-            "asset": "XCP",
-            "quantity": 92945878046,
-        },
-        {
-            "address": "mrPk7hTeZWjjSCrMTC2ET4SAUThQt7C4uK",
-            "asset": "XCP",
-            "quantity": 14999857,
-        },
-        {
-            "address": "mtQheFaSfWELRB2MyMBaiWjdDm6ux9Ezns",
-            "asset": "XCP",
-            "quantity": 99999990,
-        },
-        {
-            "address": "munimLLHjPhGeSU5rYB2HN79LJa8bRZr5b",
-            "asset": "XCP",
-            "quantity": 92999130360,
-        },
-        {
-            "address": "mwtPsLQxW9xpm7gdLmwWvJK5ABdPUVJm42",
-            "asset": "XCP",
-            "quantity": 92949122099,
-        },
-        {
-            "address": "myAtcJEHAsDLbTkai6ipWDZeeL7VkxXsiM",
-            "asset": "XCP",
-            "quantity": 92999138812,
-        },
-        {
-            "address": "tb1qw508d6qejxtdg4y5r3zarvary0c5xw7kxpjzsx",
-            "asset": "XCP",
-            "quantity": 92999030129,
-        },
-    ]
-
-
-@pytest.mark.usefixtures("api_server")
-def test_new_get_balances_vs_old():
-    asset = "XCP"
-    url = f"{config.API_ROOT}/assets/{asset}/balances"
-    new_balances = requests.get(url).json()  # noqa: S113
-    old_balance = util.api(
-        "get_balances",
-        {
-            "filters": [
-                {"field": "asset", "op": "==", "value": asset},
-                {"field": "quantity", "op": "!=", "value": 0},
-            ],
-        },
-    )
-    new_balances = sorted(new_balances, key=lambda x: (x["address"], x["asset"], x["quantity"]))
-    old_balance = sorted(old_balance, key=lambda x: (x["address"], x["asset"], x["quantity"]))
-    assert len(new_balances) == len(old_balance)
-    for new_balance, old_balance in zip(new_balances, old_balance):  # noqa: B020
-        assert new_balance["address"] == old_balance["address"]
-        assert new_balance["asset"] == old_balance["asset"]
-        assert new_balance["quantity"] == old_balance["quantity"]
-
-
-@pytest.mark.usefixtures("api_server")
-def test_new_get_asset_info():
-    asset = "NODIVISIBLE"
-    url = f"{config.API_ROOT}/assets/{asset}"
-    result = requests.get(url)  # noqa: S113
-    assert result.json() == [
-        {
-            "asset": "NODIVISIBLE",
-            "asset_longname": None,
-            "description": "No divisible asset",
-            "divisible": False,
-            "issuer": "mn6q3dS2EnDUx3bmyWc6D4szJNVGtaR7zc",
-            "locked": False,
-            "owner": "mn6q3dS2EnDUx3bmyWc6D4szJNVGtaR7zc",
-            "supply": 1000,
-        }
-    ]
-
-
-@pytest.mark.usefixtures("api_server")
-def test_new_get_asset_orders():
-    asset = "XCP"
-    url = f"{config.API_ROOT}/assets/{asset}/orders"
-    result = requests.get(url).json()  # noqa: S113
-    assert len(result) == 6
-    assert result[0] == {
-        "tx_index": 11,
-        "tx_hash": "1899b2e6ec36ba4bc9d035e6640b0a62b08c3a147c77c89183a77d9ed9081b3a",
-        "block_index": 310010,
-        "source": "mn6q3dS2EnDUx3bmyWc6D4szJNVGtaR7zc",
-        "give_asset": "XCP",
-        "give_quantity": 100000000,
-        "give_remaining": 100000000,
-        "get_asset": "BTC",
-        "get_quantity": 1000000,
-        "get_remaining": 1000000,
-        "expiration": 2000,
-        "expire_index": 312010,
-        "fee_required": 900000,
-        "fee_required_remaining": 900000,
-        "fee_provided": 6800,
-        "fee_provided_remaining": 6800,
-        "status": "open",
-    }
-
-
-@pytest.mark.usefixtures("api_server")
-def test_new_get_order_info():
-    tx_hash = "1899b2e6ec36ba4bc9d035e6640b0a62b08c3a147c77c89183a77d9ed9081b3a"
-    url = f"{config.API_ROOT}/orders/{tx_hash}"
-    result = requests.get(url).json()  # noqa: S113
-    assert result[0] == {
-        "tx_index": 11,
-        "tx_hash": "1899b2e6ec36ba4bc9d035e6640b0a62b08c3a147c77c89183a77d9ed9081b3a",
-        "block_index": 310010,
-        "source": "mn6q3dS2EnDUx3bmyWc6D4szJNVGtaR7zc",
-        "give_asset": "XCP",
-        "give_quantity": 100000000,
-        "give_remaining": 100000000,
-        "get_asset": "BTC",
-        "get_quantity": 1000000,
-        "get_remaining": 1000000,
-        "expiration": 2000,
-        "expire_index": 312010,
-        "fee_required": 900000,
-        "fee_required_remaining": 900000,
-        "fee_provided": 6800,
-        "fee_provided_remaining": 6800,
-        "status": "open",
-    }
-
-
-@pytest.mark.usefixtures("api_server")
-def test_new_get_order_matches():
-    tx_hash = "74db175c4669a3d3a59e3fcddce9e97fcd7d12c35b58ef31845a1b20a1739498"
-    url = f"{config.API_ROOT}/orders/{tx_hash}/matches"
-    result = requests.get(url).json()  # noqa: S113
-    assert result[0] == {
-        "id": "74db175c4669a3d3a59e3fcddce9e97fcd7d12c35b58ef31845a1b20a1739498_1b294dd8592e76899b1c106782e4c96e63114abd8e3fa09ab6d2d52496b5bf81",
-        "tx0_index": 492,
-        "tx0_hash": "74db175c4669a3d3a59e3fcddce9e97fcd7d12c35b58ef31845a1b20a1739498",
-        "tx0_address": "mn6q3dS2EnDUx3bmyWc6D4szJNVGtaR7zc",
-        "tx1_index": 493,
-        "tx1_hash": "1b294dd8592e76899b1c106782e4c96e63114abd8e3fa09ab6d2d52496b5bf81",
-        "tx1_address": "mtQheFaSfWELRB2MyMBaiWjdDm6ux9Ezns",
-        "forward_asset": "XCP",
-        "forward_quantity": 100000000,
-        "backward_asset": "BTC",
-        "backward_quantity": 800000,
-        "tx0_block_index": 310491,
-        "tx1_block_index": 310492,
-        "block_index": 310492,
-        "tx0_expiration": 2000,
-        "tx1_expiration": 2000,
-        "match_expire_index": 310512,
-        "fee_paid": 7200,
-        "status": "pending",
-    }
-
-
 @pytest.mark.usefixtures("server_db")
 def test_messages_table(server_db):
     cursor = server_db.cursor()
@@ -570,4 +338,3 @@ def test_messages_table(server_db):
     for row in result:
         bindings = json.loads(row["bindings"])
         assert isinstance(bindings, dict)
- """
